First step into java annotations, few issues solved.

+++ test/sintaxtree/Nodes/Operators/DivisionOperationTest.java
@@ -21,7 +21,7 @@ public class DivisionOperationTest {
     public void testDivisionIntDouble() {
         Node node = new BinaryOperation(new Constant(new IntegerType(1)),
                 new Constant(new DoubleType(1.0)), Operator.division);
-        assertEquals(IntegerType.class, node.evaluate().getClass());
+        assertEquals(DoubleType.class, node.evaluate().getClass());
         assertEquals(1.0, (double) node.evaluate().getValue(), 0.0);
     }
 
@@ -29,7 +29,7 @@ public class DivisionOperationTest {
     public void testDivisionDoubleInt() {
         Node node = new BinaryOperation(new Constant(new DoubleType(1.0)),
                 new Constant(new IntegerType(1)), Operator.division);
-        assertEquals(IntegerType.class, node.evaluate().getClass());
+        assertEquals(DoubleType.class, node.evaluate().getClass());
         assertEquals(1.0, (double) node.evaluate().getValue(), 0.0);
     }
 
@@ -37,7 +37,7 @@ public class DivisionOperationTest {
     public void testDivisionDoubleDouble() {
         Node node = new BinaryOperation(new Constant(new DoubleType(1.0)),
                 new Constant(new DoubleType(1.0)), Operator.division);
-        assertEquals(IntegerType.class, node.evaluate().getClass());
+        assertEquals(DoubleType.class, node.evaluate().getClass());
         assertEquals(1.0, (double) node.evaluate().getValue(), 0.0);
     }
 }
